backend/requirements.txt
fastapi
uvicorn
elasticsearch
python-dotenv
sentence-transformers
pydantic